+++ lighting.py
@@ -12,7 +12,7 @@
 
 
 import validators # Required to validating URLs
-import urllib.request # Required to send network requests
+import requests # Required to send network requests
 import json # Required to process JSON data
 import os # Required to interact with certain operating system functions
 
@@ -29,9 +29,6 @@ config = json.load(open(predator_root_directory + "/config.json")) # Load the co
 def update_status_lighting(url_id):
     status_lighting_update_url = str(config["realtime"]["status_lighting_values"][url_id]).replace("[U]", str(config["realtime"]["status_lighting_base_url"]))# Prepare the URL where a request will be sent in order to update the status lighting.
     if (validators.url(status_lighting_update_url)): # Check to make sure the URL ID supplied actually resolves to a valid URL in the configuration database.
-        try: # Try sending a request to the URL.
-            response = requests.get(status_lighting_update_url)
-        except Exception as e:
-            response = e
+        response = requests.get(status_lighting_update_url)
     else:
         print(style.yellow + "Warning: Unable to update status lighting. Invalid URL configured for " + url_id + style.end) # Display a warning that the URL was invalid, and no network request was sent.