+++ CHANGELOG.md
@@ -245,7 +245,12 @@ This update prefaces plans for future updates, and adds 'survey mode', which all
 - Reorganized some configuration values to make custom databases more intuitive.
     - Moved all traffic camera related configuration values to either the 'General' section or the 'Information Mode' section.
 - The second command line argumented appended to Predator can now be used to specify the project root directory.
-- Simplified enforcement camera alerts in Real-time Mode to make them appear more like enforcement camera alerts in Information Mode.
-- Enforcement camera alerts in real-time mode now support shape alerts, audio alerts, and status lighting alerts.
-- Moved the 'license plate validation format' prompt to the second prompt in Real-time Mode.
+- Reworked enforcement camera alerts in Real-time Mode to make them appear more like enforcement camera alerts in Information Mode.
+- Enforcement camera alerts in Real-time Mode now support shape alerts, audio alerts, and status lighting alerts.
+- Moved the 'license plate validation format' prompt to the second prompt in Realtime Mode.
 - The webhook and status lighting system now use the Python 'requests' library.
+- Added the ability for Information Mode to display the current speed in a large ASCII font using the `big_speed_display` configuration value.
+- In real-time mode, status lighting will now be updated before playing audio alerts.
+- Increased the default delay time on several sound effects.
+- Decreased the volume of several sound effects.
+- Added several new custom sound effects.